// back-end/src/character/character.interface.ts
export class Character {
  CHAR_SQ: number;
  CHAR_NM: string;
  CHAR_STDD: string;
  CHAR_IMG_PATH: string;
  CHAR_BGM_PATH: string;
  CHAR_DESC: string;
  CHAR_STRONG: string;
  CHAR_WEAK: string;
  CRT_DT: string;
  MOD_DT: string;
  DEL_DT: string;
  CHAR_PNT: number;
}

export class CharacterStandard {
  CHAR_STDD_SQ: number;
  CHAR_SQ: number;
  TEST_CTGR_SQ: number;
}